Shortcuts for deleting a conversation.
Closes #51
Closes #84

The works by leveraging the fact that the most recently opened popup get rearranged in the DOM and is placed last in the list.

Once the delete modal is open we can use tab selection to avoid switching to the mouse:
![image](https://cloud.githubusercontent.com/assets/3745612/14786328/5b6d48e4-0acc-11e6-9338-caa87bdef781.png)
![image](https://cloud.githubusercontent.com/assets/3745612/14786336/6a695018-0acc-11e6-9243-08619b0fc3c2.png)

The current shortcut is temporarily set to `CmdOrCtrl+Shift+K` (for Kill?). Didn't want to opt with `Cmd+Backspace` like in iMessages as that would get shadowed by the clear line command in text boxes. So suggestions are welcome.

Also added in a shortcut for mute at `CmdOrCtrl+ K`

Closes #85