Fixed back button margin
Added event for setting margin if platform is darwin, otherwise the
margin is set to 0. Also removed original style from `browser.css` for
margin because it breaks this setting. There is a problem that page
loads too slow and js returns empty HTMLCollection so I've added a
workaround for it with `setTimeout` function.