Fixed tray icon
Tray icons was not turning into `unread` state when new messages are
received. There was a bug for group chat photo because it uses
background image on div instead img tag. Also `updateBadge()` method was
inside if statement that is true only on darwin platform so tray icon
wasn't changed on linux/win32.