Made showSettingsMenu function async
showSettingsMenu is used in a cuple of places and in one place it needs
to wait for element to be created to show the menu. In that place so far
we've been using a custom function elementReady instead of
document.querySelector. Now showSettingsMenu is made async so it can now
use elementReady.