Refactor renderEmoji to async/await + better-ipc
Render emoji function was implemented with a fix for ipc calls being
answered by a single ipc response. This was using Promises. I wanted to
refactor this function to use async/await and with better-ipc which
already had implemented a fix for the mentioned problem (with uuid's for
each event) this could practically be converted into a single arrow
function.

I'd like to thank Felix Becker (@felixfbecker) for helping me out since
he originally implemented the native emoji functionality.