Monitor for changes in class list and update it
This fixes dark mode not being set on startup. Messenger website seems
to remove the `__fb-dark-mode` class just after Caprine adds it and it
disables dark mode on startup. This solution uses MutationObserver to
observe when class list in `html` element changes and if it doesn't
match the config of the app, it updates the class list. This enables
dark mode on every startup.