Fix reaction, delete and image bugs for Windows
There are some bugs on latest release that cause inability to send
reactions on messages, delete messages and conversation and load images
when clicked. This happens because hide-preferences-window class is not
removed from document when preferences window is closed because close
button couldn't be found on some locales.

It's fixed by looking for the close button with class names instead of
aria-label attribute.