Add a temporary workaround for brokend website ui
Messenger website is adding __fb-light-mode class to an element in
received messages which makes message text dark instead of light in dark
mode. This is fixed by setting the --primary-text variable to a light
color from dark mode in __fb-light-mode class only if said class is in a
child element of an element containing __fb-dark-mode class. This way,
the style is applied only in dark mode.